🔥add the option to set a custom Cache-Control value for a favicon response
Right now, each favicon response sets the `Cache-Control` header to
`public, max-age=31536000` (1 year). In some cases this could be a
little bit high if you are changing it etc. and the header is not
replaced by a reverse proxy.

With this change, the favicon middleware allows the configuration of the
`Cache-Control` header. The default behavior is to keep the setting of
`max-age=31536000 `, but you could fine-tune it.